upgrade pickables now are UpgradePickables
