+++ src/entity/upgrade_pickables/GunLevel3.java
@@ -1,25 +1,25 @@
-package entity.pickables;
+package entity.upgrade_pickables;
 
 import java.io.IOException;
 
 import javax.imageio.ImageIO;
 
-import entity.Pickable;
+import entity.UpgradePickable;
 import entity.player.Player;
 import entity.player.PlayerStats;
 import main.GameManager;
 import util.GameConstants;
 import util.Shop;
 
-public class RunnerShoes extends Pickable {
+public class GunLevel3 extends UpgradePickable {
     //#region CONSTRUCTORS
-    public RunnerShoes() {
+    public GunLevel3() {
         super();
     }
-    public RunnerShoes(int row, int col) {
+    public GunLevel3(int row, int col) {
         super(row, col);
     }
-    public RunnerShoes(double positionX, double positionY) {
+    public GunLevel3(double positionX, double positionY) {
         super(positionX, positionY);
     }
     //#endregion
@@ -27,15 +27,15 @@ public class RunnerShoes extends Pickable {
     @Override
     public void start() {
         super.start();
-        timeToLive = -1; // imperishable
+        itemCost = 15;
 
-        name = "RunnerShoes";
+        name = "GunLevel3";
     }
 
     @Override
     public void loadImages() {
         try {
-            image = ImageIO.read(getClass().getResourceAsStream("/resources/pickables/runner_shoes.png"));
+            image = ImageIO.read(getClass().getResourceAsStream("/resources/pickables/gun_level_3.png"));
         }
         catch (IOException e) {
             e.getStackTrace();
@@ -44,13 +44,14 @@ public class RunnerShoes extends Pickable {
     
     @Override
     public boolean checkPickupConditions() {
-        return (PlayerStats.removeCoins(12));
+        return (PlayerStats.removeCoins(itemCost));
     }
 
     @Override
     public void getPickedUp(Player player) {
-        GameManager.getInstance().shop.removePickableFromItemPool(Shop.RUNNER_SHOES);
-        PlayerStats.setCurrentBaseSpeed(GameConstants.Player.SPEED_RUNNER);
-        PlayerStats.setHasRunnerBoots(true);
+        super.getPickedUp(player);
+        GameManager.getInstance().shop.removePickableFromItemPool(Shop.GUN_LEVEL_UP);
+        PlayerStats.setDamage(GameConstants.Player.DAMAGE_LEVEL_3);
+        PlayerStats.setGunLevel(3);
     }
 }
